Summer2022/AutomatedLabFramework/fpga_ccu/requirements.txt
-i https://pypi.org/simple
-e .
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.7.3
pytz==2018.4
scipy==1.1.0
six==1.11.0
